feat: DF-380 - Refactor following PR
