+++ src/server/plugins/engine/components/ComponentBase.ts
@@ -9,10 +9,10 @@ import joi, {
 } from 'joi'
 
 import { type ComponentCollection } from '~/src/server/plugins/engine/components/ComponentCollection.js'
-import { type Component } from '~/src/server/plugins/engine/components/factory.js'
+import { type Component } from '~/src/server/plugins/engine/components/helpers/components.js'
 import { type ViewModel } from '~/src/server/plugins/engine/components/types.js'
 import { type FormModel } from '~/src/server/plugins/engine/models/index.js'
-import { type PageControllerClass } from '~/src/server/plugins/engine/pageControllers/factory.js'
+import { type PageControllerClass } from '~/src/server/plugins/engine/pageControllers/helpers/pages.js'
 
 export class ComponentBase {
   page?: PageControllerClass
