+++ src/server/plugins/engine/helpers.test.ts
@@ -22,7 +22,7 @@ import { buildFormContextRequest } from '~/src/server/plugins/engine/pageControl
 import {
   createPage,
   type PageControllerClass
-} from '~/src/server/plugins/engine/pageControllers/factory.js'
+} from '~/src/server/plugins/engine/pageControllers/helpers/pages.js'
 import {
   type FormContext,
   type FormContextRequest
