+++ src/server/plugins/engine/outputFormatters/machine/v1.ts
@@ -4,7 +4,7 @@ import {
 } from '@defra/forms-model'
 
 import { config } from '~/src/config/index.js'
-import { getAnswer } from '~/src/server/plugins/engine/components/factory.js'
+import { getAnswer } from '~/src/server/plugins/engine/components/helpers/components.js'
 import { FileUploadField } from '~/src/server/plugins/engine/components/index.js'
 import { type checkFormStatus } from '~/src/server/plugins/engine/helpers.js'
 import { type FormModel } from '~/src/server/plugins/engine/models/index.js'
