+++ src/server/plugins/engine/services/notifyService.ts
@@ -5,7 +5,7 @@ import {
 } from '@defra/forms-model'
 
 import { config } from '~/src/config/index.js'
-import { escapeMarkdown } from '~/src/server/plugins/engine/components/helpers.js'
+import { escapeMarkdown } from '~/src/server/plugins/engine/components/helpers/index.js'
 import { checkFormStatus } from '~/src/server/plugins/engine/helpers.js'
 import { type FormModel } from '~/src/server/plugins/engine/models/index.js'
 import { type DetailItem } from '~/src/server/plugins/engine/models/types.js'
